Move the auto-replace class to the text subfolder
